+++ backend/run_stock.sh
@@ -1,4 +1,4 @@
 #!/bin/zsh
 
 
-/usr/bin/python3 scheduler.py
+/usr/bin/python3 StockMarketCompanionApp/StockMarketCompanionApp/backend/scheduler.py